analysis-service/requirements.txt
redis
PyGithub
python-dotenv
requests
GitPython
pyright
unidiff
google-generativeai
